test1

+++ lbw/client/src/client_node.cpp
@@ -19,11 +19,11 @@ int main(int argc, char **argv) {
     n.advertise<client::Time>(srv.response.rep.c_str(), 10);
     // 设置循环的频率
     ros::Rate loop_rate(1);
-    while(ros::ok()) {
+    while(ros::ok ()) {
     // 初始化time
     ros::Time begin = ros::Time::now();
     client::Time time;
-    time.sec = (long)begin.toSec();
+    time.sec = begin.toSec();
     time.name = "client";
     // 发布消息
     client_topic_pub.publish(time);
